yolov8/requirements.txt
# YOLOv8 Tower/Antenna Classification
# Python 3.8+

# Core ML Libraries
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.23.0
pandas>=2.0.0
opencv-python>=4.8.0
Pillow>=10.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
PyYAML>=6.0
tqdm>=4.65.0
albumentations>=1.3.0

# FastAPI Server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
pydantic>=2.0.0
